backend/requirements.txt
accelerate==0.32.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
dnspython==2.4.2
fastapi==0.115.6
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.20.3
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
MarkupSafe==3.0.2
motor==3.3.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.59.7
packaging==25.0
pillow==11.2.1
psutil==7.0.0
pydantic==2.6.1
pydantic_core==2.16.2
pymongo==4.9.2
PyPDF2==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
sniffio==1.3.1
starlette==0.41.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.15.2
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.36.2
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0